// src/app/features/customers/models/customer.ts
export interface Customer {
    id: string;
    salutation: string;
    initials: string;
    firstname: string;
    firstnameAscii: string;
    gender: string;
    firstnameCountryRank: string;
    firstnameCountryFrequency: string;
    lastname: string;
    lastnameAscii: string;
    lastnameCountryRank: string;
    lastnameCountryFrequency: string;
    email: string;
    password: string;
    countryCode: string;
    countryCodeAlpha: string;
    countryName: string;
    primaryLanguageCode: string;
    primaryLanguage: string;
    balance: string;
    phoneNumber: string;
    currency: string;
}